+++ feature/login/src/main/java/com/twix/login/LoginScreen.kt
@@ -20,6 +20,7 @@ import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.twix.designsystem.R
+import com.twix.designsystem.components.button.LoginButton
 import com.twix.designsystem.components.text.AppText
 import com.twix.designsystem.components.toast.ToastManager
 import com.twix.designsystem.components.toast.model.ToastData
@@ -29,7 +30,6 @@ import com.twix.designsystem.theme.TwixTheme
 import com.twix.domain.login.LoginType
 import com.twix.domain.model.OnboardingStatus
 import com.twix.domain.model.enums.AppTextStyle
-import com.twix.login.component.LoginButton
 import com.twix.login.contract.LoginIntent
 import com.twix.login.contract.LoginSideEffect
 import com.twix.ui.base.ObserveAsEvents
